Move the mobile register link regression into the auth-flow suite
